INTYGFV-15963: Moved logic related to GetCertificateListFilterService from CertificateController to FilterController.
