+++ app/src/main/java/se/inera/intyg/minaintyg/certificate/CertificateController.java
@@ -7,11 +7,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import se.inera.intyg.minaintyg.certificate.dto.CertificateListFilterResponseDTO;
 import se.inera.intyg.minaintyg.certificate.dto.CertificateListRequestDTO;
 import se.inera.intyg.minaintyg.certificate.dto.CertificateListResponseDTO;
 import se.inera.intyg.minaintyg.certificate.dto.CertificateResponseDTO;
-import se.inera.intyg.minaintyg.certificate.service.GetCertificateListFilterService;
 import se.inera.intyg.minaintyg.certificate.service.GetCertificateService;
 import se.inera.intyg.minaintyg.certificate.service.ListCertificatesService;
 import se.inera.intyg.minaintyg.certificate.service.SendCertificateService;
@@ -25,7 +23,6 @@ import se.inera.intyg.minaintyg.certificate.service.dto.SendCertificateRequest;
 public class CertificateController {
 
   private final ListCertificatesService listCertificatesService;
-  private final GetCertificateListFilterService getCertificateListFilterService;
   private final GetCertificateService getCertificateService;
   private final SendCertificateService sendCertificateService;
 
@@ -47,7 +44,7 @@ public class CertificateController {
         .build();
   }
 
-  @PostMapping("/{certificateId}")
+  @GetMapping("/{certificateId}")
   public CertificateResponseDTO getCertificate(@PathVariable String certificateId) {
     final var response = getCertificateService.get(
         GetCertificateRequest
@@ -59,20 +56,6 @@ public class CertificateController {
     return CertificateResponseDTO
         .builder()
         .certificate(response.getCertificate())
-        .build();
-  }
-
-  @GetMapping("/filters")
-  public CertificateListFilterResponseDTO getCertificateListFilter() {
-    final var response = getCertificateListFilterService.get();
-
-    return CertificateListFilterResponseDTO
-        .builder()
-        .years(response.getYears())
-        .certificateTypes(response.getCertificateTypes())
-        .units(response.getUnits())
-        .statuses(response.getStatuses())
-        .total(response.getTotal())
         .build();
   }
